修改websocket close数据可能不存在的错误

+++ Src/WebSocket/Listener/CloseListener.php
@@ -24,13 +24,15 @@ class CloseListener extends ListenerAbstract {
 	}
 
 	private function onClose(Server $server, int $fd, int $reactorId): void {
-		$collector = icontainer()->get('ws-client')[$fd] ?? [];
-		if ($collector) {
-			/**
-			 * @var Psr7Request $psr7Request
-			 */
-			$psr7Request = $collector[0];
-			$psr7Request->session->destroy(SESSION_WEBSOCKET_CLOSE);
+		if (icontainer()->has('ws-client')) {
+			$collector = icontainer()->get('ws-client')[$fd] ?? [];
+			if ($collector) {
+				/**
+				 * @var Psr7Request $psr7Request
+				 */
+				$psr7Request = $collector[0];
+				$psr7Request->session->destroy(SESSION_WEBSOCKET_CLOSE);
+			}
 		}
 
 		//删除数据绑定记录
